app/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==3.2.0
blinker==1.9.0
cffi==1.17.1
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.3
cycler==0.12.1
databases==0.9.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
Flask==3.1.0
Flask-Cors==5.0.0
fonttools==4.55.1
greenlet==3.2.1
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.3
mysql-connector-python==9.3.0
numpy==2.1.3
opencv-python==4.10.0.84
packaging==24.2
passlib==1.7.4
pillow==11.0.0
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.10.1
PyMySQL==1.1.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-jose==3.4.0
python-multipart==0.0.20
response==0.5.0
rsa==4.9.1
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
Werkzeug==3.1.3
